requirements.txt
google-adk
python-dotenv
joblib
scikit-learn
pandas
nltk
numpy